Update Task-1- BasicGUI Calculator.py
At Codsoft, this project was undertaken to develop a simple graphical user interface (GUI) calculator using the Python programming language. The primary objective of the project was to create a user-friendly calculator that could perform basic arithmetic operations, such as addition, subtraction, multiplication, and division.

I started by designing the GUI layout, ensuring a clean and intuitive interface for users. They utilized Python's built-in GUI library, Tkinter, to implement the graphical elements of the calculator. Tkinter allowed them to create buttons for digits, operators, and the equal sign, as well as a display screen to show the input and output.

Next, I focused on the calculator's functionality. I programmed the necessary logic to handle user inputs, process arithmetic operations, and display the results accurately on the screen. Error handling mechanisms were also put in place to prevent crashes or unexpected behavior when dealing with invalid inputs.

Upon completion, the Task1 project resulted in a fully functional GUI calculator, providing a straightforward and accessible tool for users to perform basic mathematical calculations with ease.